requirements.txt
discord.py==2.3.2
selenium==4.15.2
python-dotenv==1.0.0
PyNaCl==1.5.0
